backend/pyproject.toml
[tool.poetry]
name = "ai-powered-cv-screener-backend"
version = "0.1.0"
description = "Backend API para el sistema de screening de CVs con IA"
authors = ["Tu Nombre <tu.email@ejemplo.com>"]
readme = "README.md"
packages = [{include = "backend"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
python-multipart = "^0.0.6"
pinecone-client = "^2.2.4"
openai = "^1.3.0"
python-dotenv = "^1.0.0"
pypdf2 = "^3.0.1"
langchain = "^0.0.350"
langchain-openai = "^0.0.2"
langchain-pinecone = "^0.0.3"
pandas = "^2.1.4"
numpy = "^1.25.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
